chat-api/pyproject.toml
[project]
name = "chat-api"
version = "0.7.5"
description = "Microsoft Agent Framework FoundryAgent orchestrator binding the registered submissions/tax/legal Foundry Hosted Agents (responses-protocol passthrough) with per-agent AgentSession continuity for the multi-person workflow."
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.3",
    "azure-identity>=1.19.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "pydantic>=2.11,<3",
    "agent-framework>=1.2.0",
    "pyjwt[crypto]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.21.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chat_api"]

[tool.hatch.build.targets.wheel.force-include]
"agents_prompts" = "chat_api/agents_prompts"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]
